Store models internally with a normalized name
If a swagger file uses model names that aren't starting with an uppercase letter and subclasses the code previously crashed, as the names for lookups in the models map are sometime using the original name and other times the already normalized name used for emitting the code. This change uses always a normalized name when accessing the models variable.